+++ fluent/test/functions_builtin_test.js
@@ -21,7 +21,7 @@ suite('Built-in functions', function() {
     });
 
     test('formats the number', function() {
-      const msg = ctx.messages.get('foo');
+      const msg = ctx.getMessage('foo');
       const val = ctx.format(msg, args, errs);
       assert.equal(val, '1');
       assert.equal(errs.length, 0);
@@ -41,7 +41,7 @@ suite('Built-in functions', function() {
 
     test('formats the date', function() {
       const date = new Date('2016-09-29');
-      const msg = ctx.messages.get('foo');
+      const msg = ctx.getMessage('foo');
       const val = ctx.format(msg, { date }, errs);
       // format the date argument to account for the testrunner's timezone
       assert.equal(val, dtf.format(date));