+++ fluent/test/functions_test.js
@@ -21,7 +21,7 @@ suite('Functions', function() {
     });
 
     test('falls back to the name of the function', function() {
-      const msg = ctx.messages.get('foo');
+      const msg = ctx.getMessage('foo');
       const val = ctx.format(msg, args, errs);
       assert.equal(val, 'MISSING()');
       assert.equal(errs.length, 1);
@@ -53,7 +53,7 @@ suite('Functions', function() {
     // XXX Gracefully handle wrong argument types passed into FTL Functions
     // https://bugzil.la/1307124
     it.skip('falls back when arguments don\'t match the arity', function() {
-      const msg = ctx.messages.get('pass-nothing');
+      const msg = ctx.getMessage('pass-nothing');
       const val = ctx.format(msg, args, errs);
       assert.equal(val, 'IDENTITY()');
       assert.equal(errs.length, 1);
@@ -61,21 +61,21 @@ suite('Functions', function() {
     });
 
     test('accepts strings', function() {
-      const msg = ctx.messages.get('pass-string');
+      const msg = ctx.getMessage('pass-string');
       const val = ctx.format(msg, args, errs);
       assert.equal(val, 'a');
       assert.equal(errs.length, 0);
     });
 
     test('accepts numbers', function() {
-      const msg = ctx.messages.get('pass-number');
+      const msg = ctx.getMessage('pass-number');
       const val = ctx.format(msg, args, errs);
       assert.equal(val, '1');
       assert.equal(errs.length, 0);
     });
 
     test('accepts entities', function() {
-      const msg = ctx.messages.get('pass-message');
+      const msg = ctx.getMessage('pass-message');
       const val = ctx.format(msg, args, errs);
       assert.equal(val, 'Foo');
       assert.equal(errs.length, 0);
@@ -84,21 +84,21 @@ suite('Functions', function() {
     // XXX Accept complex types (e.g. attributes) as arguments to FTL Functions
     // https://bugzil.la/1307120
     it.skip('accepts attributes', function() {
-      const msg = ctx.messages.get('pass-attr');
+      const msg = ctx.getMessage('pass-attr');
       const val = ctx.format(msg, args, errs);
       assert.equal(val, 'Attribute');
       assert.equal(errs.length, 0);
     });
 
     test('accepts externals', function() {
-      const msg = ctx.messages.get('pass-external');
+      const msg = ctx.getMessage('pass-external');
       const val = ctx.format(msg, { ext: "Ext" }, errs);
       assert.equal(val, 'Ext');
       assert.equal(errs.length, 0);
     });
 
     test('accepts function calls', function() {
-      const msg = ctx.messages.get('pass-function-call');
+      const msg = ctx.getMessage('pass-function-call');
       const val = ctx.format(msg, args, errs);
       assert.equal(val, '1');
       assert.equal(errs.length, 0);