+++ fluent/test/tags_test.js
@@ -52,28 +52,28 @@ suite('Tags', function() {
   });
 
   test('match in a message', function() {
-    const msg = ctx.messages.get('ref-foo');
+    const msg = ctx.getMessage('ref-foo');
     const val = ctx.format(msg, args, errs);
     assert.equal(val, 'Foo Tag');
     assert.equal(errs.length, 0);
   });
 
   test('match one of two', function() {
-    const msg = ctx.messages.get('ref-bar-1');
+    const msg = ctx.getMessage('ref-bar-1');
     const val = ctx.format(msg, args, errs);
     assert.equal(val, 'Bar Tag 1');
     assert.equal(errs.length, 0);
   });
 
   test('match in order of variants', function() {
-    const msg = ctx.messages.get('ref-bar-2');
+    const msg = ctx.getMessage('ref-bar-2');
     const val = ctx.format(msg, args, errs);
     assert.equal(val, 'Bar Tag 2');
     assert.equal(errs.length, 0);
   });
 
   test('fallback when tag is missing', function() {
-    const msg = ctx.messages.get('ref-baz');
+    const msg = ctx.getMessage('ref-baz');
     const val = ctx.format(msg, args, errs);
     assert.equal(val, 'Other');
     assert.equal(errs.length, 1);